backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
web3>=6.11.0
numpy>=1.24.0
scikit-learn>=1.3.0
